// src/bahanbaku/dto/create-bahanbaku.dto.ts
import { IsEmpty, MinLength } from "class-validator";

export class CreateBahanbakuDto {
    id: string;
    name: string;
    id_Bahanbaku: string;
    stock: number;
    is_stock: number;
    harga: number;
    is_minus_stock: number;
    qty_plus_minus: number;
    exp_date: Date;
    createdAt: Date;
    updatedAt: Date;
}